requirements.txt
azure-identity==1.15.0
requests